requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
pdfminer.six==20231228
python-docx==1.1.2
groq
gunicorn
requests
